+++ src/Service/Loading/Command/GetFullscreenLoadingElementCommand.mjs
@@ -1,0 +1,47 @@
+import { FullscreenLoadingElement } from "../../../Adapter/Loading/FullscreenLoadingElement.mjs";
+
+/** @typedef {import("../../../../../flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
+/** @typedef {import("../Port/LoadingService.mjs").LoadingService} LoadingService */
+
+export class GetFullscreenLoadingElementCommand {
+    /**
+     * @type {CssApi}
+     */
+    #css_api;
+    /**
+     * @type {LoadingService}
+     */
+    #loading_service;
+
+    /**
+     * @param {CssApi} css_api
+     * @param {LoadingService} loading_service
+     * @returns {GetFullscreenLoadingElementCommand}
+     */
+    static new(css_api, loading_service) {
+        return new this(
+            css_api,
+            loading_service
+        );
+    }
+
+    /**
+     * @param {CssApi} css_api
+     * @param {LoadingService} loading_service
+     * @private
+     */
+    constructor(css_api, loading_service) {
+        this.#css_api = css_api;
+        this.#loading_service = loading_service;
+    }
+
+    /**
+     * @returns {FullscreenLoadingElement}
+     */
+    getFullscreenLoadingElement() {
+        return FullscreenLoadingElement.new(
+            this.#css_api,
+            this.#loading_service.getLoadingElement()
+        );
+    }
+}