+++ src/Service/Loading/Port/LoadingService.mjs
@@ -1,6 +1,8 @@
+import { GetFullscreenLoadingElementCommand } from "../Command/GetFullscreenLoadingElementCommand.mjs";
 import { GetLoadingElementCommand } from "../Command/GetLoadingElementCommand.mjs";
 
 /** @typedef {import("../../../../../flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
+/** @typedef {import("../../../Adapter/Loading/FullscreenLoadingElement.mjs").FullscreenLoadingElement} FullscreenLoadingElement */
 /** @typedef {import("../../../Adapter/Loading/LoadingElement.mjs").LoadingElement} LoadingElement */
 
 export class LoadingService {
@@ -25,6 +27,17 @@ export class LoadingService {
      */
     constructor(css_api) {
         this.#css_api = css_api;
+    }
+
+    /**
+     * @returns {FullscreenLoadingElement}
+     */
+    getFullscreenLoadingElement() {
+        return GetFullscreenLoadingElementCommand.new(
+            this.#css_api,
+            this
+        )
+            .getFullscreenLoadingElement();
     }
 
     /**